refactor : use dynamic route
