+++ src/pages/mypage/body/answers.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { observer } from 'mobx-react-lite';
+
+const Answers = () => {
+  return (
+    <p>answers</p>
+  );
+};
+
+export default observer(Answers);
