+++ src/pages/mypage/body/index.js
@@ -1,0 +1,4 @@
+export { default as Answers } from './answers.js';
+export { default as Bookmark } from './bookmark.js';
+export { default as Profile } from './profile.js';
+export { default as Qeustions } from './questions.js';
