+++ src/pages/mypage/body/questions.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { observer } from 'mobx-react-lite';
+
+const Questions = () => {
+  return (
+    <p>questions</p>
+  );
+};
+
+export default observer(Questions);
